requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
PyPDF2==3.0.1
pdfplumber==0.10.0
pymupdf==1.23.26
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
openai==1.6.1
tiktoken==0.5.2
chromadb==0.4.18
faiss-cpu==1.7.4
sentence-transformers==2.2.2
transformers==4.36.2
torch==2.1.2
numpy==1.24.4
pandas==2.1.4
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
requests==2.31.0
aiofiles==23.2.0
tqdm==4.66.1
loguru==0.7.2
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2 
